ML_Case_6/requirements.txt
fastapi
pandas
uvicorn
sqlalchemy
pyjwt
passlib
python-jose
python-multipart
jinja2